ethereum: Move all of the Ethereum components to their new home.
This is being done to:

 * Make deployment/testing easier when contracts have dependencies on
   other contracts, by moving things into a single truffle workspace.
 * Avoid splattering web3 all over the place.

Every backend component that incorporates an Ethereum component should
live under this new `ethereum` top level crate going forward.
