+++ ethereum/src/beacon.rs
