+++ ethereum/src/lib.rs
@@ -1,13 +1,15 @@
-// ! Ekidem Ethereum random beacon backend.
+// ! Ekiden Ethereum backend components.
 extern crate chrono;
 extern crate ekiden_beacon_base;
 extern crate ekiden_common;
 extern crate ethabi;
+extern crate rustc_hex;
 extern crate serde_json;
 extern crate web3;
 #[macro_use]
 extern crate log;
 
-mod backend;
+mod beacon;
+pub mod truffle;
 
-pub use backend::EthereumRandomBeacon;
+pub use beacon::EthereumRandomBeacon;
