+++ ethereum/src/truffle.rs
@@ -1,7 +1,11 @@
+use ekiden_common::error::Error;
+use ekiden_common::futures::{future, stream, BoxStream, Future};
 use rustc_hex::FromHex;
-use std::env;
+use std;
 use std::io::Read;
 use std::process::{Child, Command, Stdio};
+use std::{env, thread, time};
+use web3;
 
 /// The hard coded truffle development Ethereum address, taken from
 /// http://truffleframework.com/docs/getting_started/console (entry 0).
@@ -67,4 +71,19 @@ pub fn test_truffle(cwd: &str) {
         .status()
         .expect("truffle failed");
     assert!(status.success());
+}
+
+/// Make a stream of transactions between two truffle default accts to keep the chain going.
+pub fn mine<T: 'static + web3::Transport + Sync + Send>(tport: T) -> BoxStream<u64>
+where
+    <T as web3::Transport>::Out: std::marker::Send,
+{
+    Box::new(stream::unfold(0, move |state| {
+        thread::sleep(time::Duration::from_millis(500));
+        Some(
+            tport
+                .execute("evm_mine", vec![])
+                .then(move |_r| future::ok::<(u64, u64), Error>((0, state + 1))),
+        )
+    }))
 }
