+++ ethereum/test/beacon.js
